# swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ocp-remind-api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "RemindApiV1"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/create": {
      "post": {
        "operationId": "RemindApiV1_CreateRemind",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiRemind"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiCreateRemindRequest"
            }
          }
        ],
        "tags": [
          "RemindApiV1"
        ]
      }
    },
    "/v1/describe/{RemindId}": {
      "get": {
        "operationId": "RemindApiV1_DescribeRemind",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiRemind"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "RemindId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "RemindApiV1"
        ]
      }
    },
    "/v1/list": {
      "post": {
        "operationId": "RemindApiV1_ListReminds",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiListRemindsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "RemindApiV1"
        ]
      }
    },
    "/v1/remove": {
      "delete": {
        "operationId": "RemindApiV1_RemoveRemind",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiRemoveRemindResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "Id",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "RemindApiV1"
        ]
      }
    }
  },
  "definitions": {
    "apiCreateRemindRequest": {
      "type": "object",
      "properties": {
        "UserId": {
          "type": "string",
          "format": "uint64"
        },
        "Text": {
          "type": "string"
        },
        "time": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "apiListRemindsResponse": {
      "type": "object",
      "properties": {
        "Ids": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "uint64"
          }
        }
      }
    },
    "apiRemind": {
      "type": "object",
      "properties": {
        "Id": {
          "type": "string",
          "format": "uint64"
        },
        "UserId": {
          "type": "string",
          "format": "uint64"
        },
        "Text": {
          "type": "string"
        },
        "UnixTime": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "apiRemoveRemindResponse": {
      "type": "object",
      "properties": {
        "Result": {
          "type": "boolean"
        },
        "ErrMsg": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "typeUrl": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}